T2 - Refactor - extract userbadge to component
